requirements.txt
streamlit==1.31.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
beautifulsoup4==4.12.3
tld==0.13
python-whois==0.8.0
shap==0.44.0
joblib==1.3.2
plotly==5.18.0 